add page to collection in Update _config.yml
Hey @mahdi-robbani and @nkurtys , 

I've noticed that the host-isolate thingy did not pop on on the dev and main pages.
This commit adds it to the config.yml file. 

Possible issues: 
- the page has been directly merged to main (not dev) > do you know a quick solve?

Resolved issues: I've merged a PR regarding the refactoring pages and how to add this to the contribution file. I'll add an extra note in the contribution doc and mergeg this one, too.